# backend/src/API.yml
openapi: 3.0.3
info:
  description: This is the interactive documentation for the Innovatiehuis Intranet RESTful API
  version: "0.0.0"
  title: Innovatieplatform REST API
servers:
  - url: http://127.0.0.1:5000/api



  # Paths supported by the server application
paths:
...
  /projects/{project_id}/folders:
    get:
      operationId: src.endpoints.folders.get_folders_in_path
      tags:
        - Folders
      summary: Returns folder(s) in directory
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/path'
      responses:
        200:
          description: Successfully retrieved folders at given path
    put:
      operationId: src.endpoints.folders.update_dir
      tags:
        - Folders
      summary: Move folder to another directory
      parameters:
        - $ref: '#/components/parameters/project_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                to:
                  type: string
                rename:
                  type: string
      responses:
        200:
          description: Successfully moved folder
        400:
          description: Failed to move folder
    delete:
      operationId: src.endpoints.folders.delete_dir
      tags:
        - Folders
      summary: Delete folder and file(s) in directory
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/path'
        - $ref: '#/components/parameters/conf'
      responses:
        200:
          description: Succesfully delete folder
        400:
          description: Failed to delete folder
        409: 
          description: There are files or folders in this folder, requires confirmation
    post:
      operationId: src.endpoints.folders.create_dir
      tags:
        - Folders
      summary: Place new folder directory
      parameters:
        - $ref: '#/components/parameters/project_id'
        - $ref: '#/components/parameters/path'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        200:
          description: Succesfully add new folder


# ================================ANNOUNCEMENTS===================================
...
components:
  parameters:
    project_id:
      in: path
      name: project_id
      required: true
      schema:
        type: string
      description: The ID number of a project
...
    path:
      in: query
      name: path
      required: true
      schema:
        type: string
      description: The path to be used
    conf:
      in: query
      name: conf
      required: false
      schema:
        type: string
      description: Confirmation for an action
...